requirements.txt
django==2.0

djangorestframework

pymsql

requests






